README: Fix cljsbuild example
The previous example had a format that cljsbuild can't understand, so
the test build configuration was not being taken into account by
cljsbuild.

Cljsbuild expects build configurations to be under the `:builds` entry.
It even still supports (but shows a deprecation message) having the
build options right as first level entries in the cljsbuild map, but our
previous version had `:test` as first level and the build options
inside it.