Permit (:builds (:cljsbuild project)) to be a list.
When the builds collection is part of a composite profile[1], it may be
a list instead of a vector. This is probably a bug in Leiningen, and I
have reported it[2], but this commit provides a trivial workaround:
apply `mapv` to the builds collection whether it is a vector or a seq.
The result becomes a vector.

[1] https://github.com/technomancy/leiningen/blob/master/doc/PROFILES.md#composite-profiles
[2] https://github.com/technomancy/leiningen/issues/2119